Add ts-macros and logIfNotProduction macro
Note that ts-macros requires compiler support, which I wasn't yet able
to get working with create-react-app. So macros do not yet work.

Howeve, the ts-macros author helpfully created an example of how to get
it working in Vite, so we can use this to get macros working after we
switch to Vite.
https://github.com/GoogleFeud/ts-macros-vite-example/tree/main/src