Add example code of overriding gas parameters
This is an example of how to override maxFeePerGas and
maxPriorityFeePerGas in usePrepareContractWrite. It includes handling of
some important edge cases, such as maxFeePerGas being too low,
maxFeePerGas being insufficiently greater than the lastBaseFeePerGas,
and maxFeePerGas being less than maxPriorityFeePerGas. It also shows how
to set maxFeePerGas and maxPriorityFeePerGas based on useFeeData; and
how to multiply BigNumbers by floats (ie. multiply and then divide by
integral numbers representing the fractionalized float).

However, this effort to set maxFeePerGas and maxPriorityFeePerGas was
ultimately unsuccessful because it seems to be very tricky to set these
values in a multichain context. For example, arbitrum uses a different
"gas scale" than the L1 and some other L2s, such that an erc20 transfer
on arb takes over 2 million gas, but about 90% less on the L1 and some
other L2s, and so proper values for arbitrum's (and optimism's)
maxFeePerGas tends to be orders of magnitudes less than the L1. And yet,
if one simply sets maxFeePerGas based on the lastBaseFeePerGas, for some
chains, this will result in an error state where, for example, on linea
testnet the lastBaseFeePerGas may be reported as, say, 12 wei (not
gwei), but if you submit a transaction to linea testnet with
maxBaseFee=12 wei, it'll error. So then we tried to start enforcing a
minimum maxBaseFee, but again this is difficult as the appropriate
minimum for arbitrum and optimism is too small on linea testnet.

In short, setting maxFeePerGas and maxPriorityFeePerGas is best left to
wagmi+wallets (which is what happens by default if you don't override
these params), and so we'll revert this code and it's only being
committed as an example.