Add HideFooter.tsx
HideFooter is a library for route-based control of hiding the footer
in contexts that would otherwise include a footer by default.

For example, in a route hiearchy like

HideFooterOnMobile
  MainWrapper
    RequestMoney

The ancestor route of HideFooterOnMobile is detected by MainWrapper to
suppress the footer on mobile. This allows RequestMoney to leverage
MainWrapper while meeting its product-specific need to show no footer on
mobile.