Fix Web Share data
In Pay and Request, we were previously including only ShareData.text and
excluding ShareData.url. But, in some platforms like Android progressive
web app, the text is ignored if the url is not present (and maybe text
ignored even if url is present) and so this results in the share being a
no-op where you're sharing no content, and the request can't be shared
besides scanning the QR code. So this is a bugfix to always include the
request URL for Web Share platforms that require it. Unfortunately, if
both ShareData.url/text are provided, the text will be ignored on some
platforms, leading to loss of helpful context (eg. loss of "Hey please
pay me $5") and since we don't yet have URL previews, the link itself
doesn't provide any context on the payment details.