Add multi currency pay links
It's now possible to request $5 and get paid $5 in ETH at the current
spot price.

Added CheckoutSettings protobuf checkout_settings_major_version.

useBestStrategy was revamped to support the rapid strategies
regeneration caused by exchange rate updates. There was an existing
state sync bug that was exacerbated by the rapid regeneration.

Exchange rate conversions and decimals are fairly dangerous right now
because the client needs to manually keep track of both the denomination
and the decimals. In the future, it should become automatic with
typesafey.

We now use Uppercase<string> for all token and native currency tickers
in an effort to prevent case sensitivity bugs at the type level.