Revert "Set client status undefined on unmount in ExecuteTokenTransferButton"
This reverts commit 82d509dae19e0f2e45d4c8ce9acde17f40722623.

This commit is correct in its approach to set client status to undefined
on ExecuteTokenTransferButton unmounted because a client should never
depend on a dangling button status.

However, our Pay currently does depend on a dangling status: the final
button status before the button is unmounted is currently used to drive
the post-payment success UX in Pay.

This commit should be re-reverted in conjunction with a change to split
the post-payment UX into a new route that guarantees that payment can't
happen twice (since the pay route has been left) and simplifies success
state tracking since this entire new page is dedicated to post-payment
UX.