Add Token/NativeCurrency.tickerCanonical
tickerCanonical is our facility to allow a token's internal system
ticker (ie. nativeCurrency/token.ticker) to be different than its
canonical ticker that is used in renderers to display to end-users. This
allows us to support both native and bridged USDC on the same chain (and
other similar tokens in the future).

WARNING TokenKey for Tokens is no longer unique per (ticker, chain) and
has been updated to use token.contractAddress. System must not depend on
the uniqueness of (ticker, chain) for tokens. But, the uniqueness of
(ticker, chain) still holds for native currencies, and native currency
TokenKeys are still generated using this pair.

See design note that explains why all bridged and native variants of
USDC share the same ticker "USDC".

Renderers were updated to prefer tickerCanonical over ticker. Some
renderers that depended only on a ticker had to be updated to depend
instead on the higher-level Token or NativeCurrency. This is probably
the natural dependency and how it should have worked already.

Add ArbitrumUSDCNative.