Enable merchant checkout as an iframe modal pop-up
This change allows 3cities to be used to power merchant checkouts via
3cities being embedded as an iframe modal pop-up.

An expanded checkout success action API, eg. may be used to auto-close
the iframe pop-up upon successful checkout.

When running in an iframe, 3cities will use iframe message passing to
securely communicate checkout lifecycle events with the parent window.
This allows the client's webpage to receive 3cities payment details.
Currently, we support two kinds of messages, 'Checkout' (upon successful
checkout) and 'CloseIframe' (to request that the parent window close the
3cities iframe).

Optionally require 3cities to be running in an iframe for checkout to
proceed. This can help avoid users submitting payments in standalone
windows when client recognition of those payments depends on iframe
message passing.

Optional sender/buyer address authentication, so that the sender/buyer
must sign a message to prove ownership of their address prior to
checking out. When running in an iframe, the signature is provided to
the client via iframe message passing upon sucessful checkout. Clients
may use this signature to securely associate an onchain transaction with
the client's view of the checkout.

The 3cities pay link URL param API was updated support some of the new
API surface. There's no real SDK right now, so clients must use the URL
param API to use 3cities as an iframe.