Remove Web3Auth
We developed a Web3auth (oauth wallet as a service company) connector in
Q1 2023 with the expectation that this would be the best way to onboard
new sellers who didn't yet have wallets to receive payments using
3cities. However, due to limitations discussed in the diff, we never
shipped this. Today in Q2 2024, our preferred way to support new sellers
without wallets is likely Coinbase Smart Wallet. So, we're removing
Web3Auth entirely.