Add 'change wallet' button
The 'change wallet' button just disconnects the current wallet, and then
a new wallet may be connected.

Some customers did not realize that disconnecting/changing the wallet
was already possible by clicking on their address in top right and then
clicking 'disconnect' in the connectkit modal. This new button should
make it clearer.