Suggest to sender how they might afford a payment
If the sender can't afford to pay, 3cities can't allow payment to
proceed and gives an "insufficient funds" warning to the user.

However, some users were confused by this warning when they had some
wallet balances that were partially or nearly sufficient to complete the
payment.

3cities now provides the sender with a breakdown of their account
balances that are closest to being able to afford payment, and
convenient deposit links to top up the minimum necessary added balance
to complete the payment.